+++ examples/github-webhook/src/index.ts
@@ -18,6 +18,22 @@ new Trigger({
 }).listen();
 
 new Trigger({
+  id: "github-issue-comment",
+  name: "GitHub issue comment: triggerdotdev/trigger.dev-examples",
+  apiKey: "trigger_dev_zC25mKNn6c0q",
+  endpoint: "ws://localhost:8889/ws",
+  logLevel: "debug",
+  on: github.events.issueCommentEvent({
+    repo: "triggerdotdev/trigger.dev-examples",
+  }),
+  run: async (event, ctx) => {
+    await ctx.logger.info(`Action was ${event.action}`);
+
+    return {};
+  },
+}).listen();
+
+new Trigger({
   id: "github-webhook-pull_request",
   name: "GitHub PR: triggerdotdev/trigger.dev-examples",
   apiKey: "trigger_dev_zC25mKNn6c0q",