+++ packages/internal-integrations/src/shopify/index.ts
@@ -1,13 +1,13 @@
+import { shopify } from "@trigger.dev/providers";
+import { Client, createClient } from "@urql/core";
 import debug from "debug";
+import type { ReactNode } from "react";
 import {
   DisplayProperties,
   PerformedRequestResponse,
   PerformRequestOptions,
   RequestIntegration,
 } from "../types";
-import { Client, createClient, gql } from "@urql/core";
-import { shopify } from "@trigger.dev/providers";
-import { z } from "zod";
 import {
   addProductsToCollectionQuery,
   appendProductImagesQuery,
@@ -165,6 +165,10 @@ class ShopifyRequestIntegration implements RequestIntegration {
         };
       }
     }
+  }
+
+  renderComponent(input: any, output: any): ReactNode {
+    return null;
   }
 
   async #searchProductVariants(