+++ packages/trigger-integrations/src/index.ts
@@ -1,5 +1,6 @@
 import * as github from "./integrations/github";
+import * as resend from "./integrations/resend";
 import * as slack from "./integrations/slack";
 import * as shopify from "./integrations/shopify";
 
-export { github, slack, shopify };
+export { github, resend, slack, shopify };